cargos_backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-crispy-forms==1.8.1
django-dbbackup==3.2.0
django-dbbackup-ui==0.2.1
django-heroku==0.3.1
django-model-utils==3.2.0
django-notifications-hq==1.5.0
django-suit==2.0a1
django-tempus-dominus==5.1.2.9
djangorestframework==3.11.0
get==2019.4.13
gunicorn==19.9.0
idna==2.8
jsonfield==2.0.2
Pillow==6.2.1
post==2019.4.13
psycopg2==2.8.3
public==2019.4.13
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
